third_party/blink/web_tests/platform/generic/fast/files/url-required-arguments-expected.txt

Test required arguments of URL object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS URL.createObjectURL() threw exception TypeError: Failed to execute 'createObjectURL' on 'URL': 1 argument required, but only 0 present..
PASS URL.revokeObjectURL() threw exception TypeError: Failed to execute 'revokeObjectURL' on 'URL': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
